pyproject.toml
[project]
name = "python-project-template"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno>=1.2.4",
    "duckduckgo-search>=7.5.4",
    "lancedb>=0.21.2",
    "lxml-html-clean>=0.4.1",
    "newspaper4k>=0.9.3.1",
    "numpy>=2.2.4",
    "openai>=1.68.2",
    "pandas>=2.2.3",
    "pylance>=0.25.1",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.39",
    "tantivy>=0.22.2",
    "yfinance>=0.2.55",
]

[tool.mypy]
strict = true

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D", "ANN", "TRY", "EM", "FA", "SIM108", "S101", "RET505", "TCH002"]
